+++ package.json
@@ -23,11 +23,9 @@
   },
   "devDependencies": {
     "cli": "^1.0.1",
-    "doxdox": "^2.0.3",
-    "ink-docstrap": "^1.3.2",
     "jsdoc": "^3.5.5",
+    "jsdock": "^1.0.4",
     "markdox": "^0.1.10",
-    "minami": "^1.2.3",
     "mocha": "^4.1.0",
     "standard": "^10.0.3",
     "steam-dummy": "^0.0.1"
@@ -36,10 +34,11 @@
     "lint": "yarn standard",
     "test": "mocha test",
     "full-test": "yarn lint && yarn test",
-    "clean-docs": "rm -rf ./docs && rm ./API.md",
-    "jsdoc": "yarn clean-docs && jsdoc -c .jsdoc.json -R README.md",
-    "jsdoc-strap": "yarn clean-docs && jsdoc -c .jsdoc.json -R README.md -t ./node_modules/ink-docstrap/template",
-    "jsdoc-minami": "yarn clean-docs && jsdoc -c .jsdoc.json -R README.md -t ./node_modules/minami",
-    "jsdoc-dox": "jsdox ./steamconfig.js -o ./docs"
+    "jsdoc": "jsdoc -c .jsdoc.json -R README.md",
+    "jsdoc-strap": "jsdoc -c .jsdoc.json -R README.md -t ./node_modules/ink-docstrap/template",
+    "markdox": "markdox src/steamconfig.js -o API.md",
+    "fixmd": "node fix_md_async.js -c steamConfig -m ./API.md -j ./src/steamconfig.js",
+    "jsdock-build": "jsdock clean && jsdock build",
+    "jsdock-publish": "jsdock build && jsdock publish"
   }
 }
